# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/books/{id}:
    get:
      tags:
      - api-book-controller
      operationId: getBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - api-book-controller
      operationId: editBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: author
        in: query
        required: false
        schema:
          type: string
      - name: ISBN
        in: query
        required: false
        schema:
          type: string
      - name: pages
        in: query
        required: false
        schema:
          type: string
      - name: genre
        in: query
        required: false
        schema:
          type: string
      - name: date
        in: query
        required: false
        schema:
          type: string
      - name: publisher
        in: query
        required: false
        schema:
          type: string
      - name: series
        in: query
        required: false
        schema:
          type: string
      - name: description
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - api-book-controller
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/reviews:
    get:
      tags:
      - api-review-controller
      summary: Get reviews for a specific book
      operationId: getReviews
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: book
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "400":
          description: Invalid book ID
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Reviews found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_BasicInfo'
        "404":
          description: Book not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - api-review-controller
      summary: Add a review for a specific book
      operationId: addReview
      parameters:
      - name: book
        in: query
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review_ReviewBasicView'
        required: true
      responses:
        "400":
          description: Invalid book ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewBasicView'
        "201":
          description: Review added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_ReviewBasicView'
        "404":
          description: Book not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewBasicView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewBasicView'
...
  /api/auth/refresh:
    post:
      tags:
      - api-login-controller
      summary: Refresh token
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "400":
          description: Invalid token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/auth/login:
    post:
      tags:
      - api-login-controller
      summary: Login
      operationId: login
      parameters:
      - name: accessToken
        in: cookie
        required: false
        schema:
          type: string
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Invalid credentials
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/reviews/{reviewID}:
    get:
      tags:
      - api-review-controller
      summary: Get a review
      operationId: getReview
      parameters:
      - name: reviewID
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid review ID
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Review found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_BasicInfo'
    delete:
      tags:
      - api-review-controller
      summary: Delete a review
      operationId: deleteReview
      parameters:
      - name: reviewID
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Review deleted
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid review ID
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    Book_ReviewBasicView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isbn:
          type: string
    Review_ReviewBasicView:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        title:
          type: string
        authorName:
          type: string
        rating:
          type: integer
          format: int32
        content:
          type: string
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User_ReviewBasicView'
    User_ReviewBasicView:
      type: object
      properties:
        readBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_ReviewBasicView'
        readingBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_ReviewBasicView'
        wantedBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_ReviewBasicView'
        role:
          type: array
          items:
            type: string
    Book_BasicInfo:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        title:
          type: string
        authorString:
          type: string
        description:
          type: string
        releaseDate:
          type: string
        ISBN:
          type: string
        averageRating:
          type: number
          format: double
        series:
          type: string
        pageCount:
          type: integer
          format: int32
        publisher:
          type: string
        id:
          type: integer
          format: int64
        isbn:
          type: string
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
    User_BasicInfo:
      type: object
      properties:
        readBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        readingBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        wantedBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        role:
          type: array
          items:
            type: string
    Review_BasicInfo:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        title:
          type: string
        authorName:
          type: string
        rating:
          type: integer
          format: int32
        content:
          type: string
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User_BasicInfo'
...